backend/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
python-docx==0.8.11
reportlab
PyPDF2==3.0.1
python-multipart